+++ packages/atomic/src/components/commerce/atomic-product-section-actions/atomic-product-section-actions.new.stories.tsx
@@ -10,18 +10,17 @@ const {events, args, argTypes, template} = getStorybookHelpers(
   {excludeCategories: ['methods']}
 );
 
-const {decorator: commerceInterfaceDecorator, afterEach} =
-  wrapInCommerceInterface({
-    engineConfig: {
-      preprocessRequest: (request) => {
-        const parsed = JSON.parse(request.body as string);
-        parsed.perPage = 1;
-        request.body = JSON.stringify(parsed);
-        return request;
-      },
+const {decorator: commerceInterfaceDecorator, play} = wrapInCommerceInterface({
+  engineConfig: {
+    preprocessRequest: (request) => {
+      const parsed = JSON.parse(request.body as string);
+      parsed.perPage = 1;
+      request.body = JSON.stringify(parsed);
+      return request;
     },
-    includeCodeRoot: false,
-  });
+  },
+  includeCodeRoot: false,
+});
 
 const {decorator: commerceProductListDecorator} = wrapInCommerceProductList(
   'grid',
@@ -54,7 +53,7 @@ export const Default: Story = {
     commerceProductListDecorator,
     commerceInterfaceDecorator,
   ],
-  afterEach,
+  play,
   args: {
     'default-slot': `<button class="btn btn-primary">Add to Cart</button>`,
   },