+++ packages/atomic/src/components/commerce/product-list/product-template-provider.spec.ts
@@ -3,7 +3,7 @@ import {
   type Product,
 } from '@coveo/headless/commerce';
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest';
-import type {ItemTarget} from '@/src/components/common/layout/display-options';
+import type {ItemTarget} from '@/src/components/common/layout/item-layout-utils';
 import type {TemplateProviderProps} from '@/src/components/common/template-provider/template-provider';
 import {ProductTemplateProvider} from './product-template-provider';