+++ packages/atomic/src/components/search/result-template-components/atomic-result-image/atomic-result-image.new.stories.tsx
@@ -20,8 +20,9 @@ const {decorator: resultDecorator, engineConfig} = wrapInResult({
     },
   },
 });
-const {decorator: searchInterfaceDecorator, afterEach} =
-  wrapInSearchInterface(engineConfig);
+const {decorator: searchInterfaceDecorator, play} = wrapInSearchInterface({
+  config: engineConfig,
+});
 
 const meta: Meta = {
   component: 'atomic-result-image',
@@ -38,7 +39,7 @@ const meta: Meta = {
   args,
   argTypes,
 
-  afterEach,
+  play,
 };
 
 export default meta;